+++ app/lib/screens/code/constants.ts
@@ -1,0 +1,12 @@
+/**
+ * default message
+ * @todo - detect the platform and customize the message.
+ */
+export const DEFAULT_EMPTY_CODE_SNIPPET = `//
+ //
+ //
+ // there is no selected design.
+ // select your screen or component on figma
+ //
+ //
+ //`;