+++ figma/src/code-thread/index.ts
@@ -1,0 +1,2 @@
+export * from "./message-handler";
+export * from "./selection";